refactor(server): extract command workspace scope
