+++ server/proliferate/server/cloud/commands/service.py
@@ -11,15 +11,10 @@ from proliferate.constants.cloud import (
     CloudCommandActorKind,
     CloudCommandKind,
     CloudTargetStatus,
-    CloudWorkspaceStatus,
 )
 from proliferate.db.session_ops import is_integrity_error
-from proliferate.db.store import cloud_runtime_environments, cloud_workspaces
 from proliferate.db.store.cloud_sync import commands as commands_store
-from proliferate.db.store.cloud_sync import events as events_store
-from proliferate.db.store.cloud_sync import exposures as exposures_store
 from proliferate.db.store.cloud_sync import targets as targets_store
-from proliferate.server.cloud.claims.access import require_workspace_interact
 from proliferate.server.cloud.commands.client_state import (
     expire_stale_client_command_if_needed,
     record_pending_prompt_interaction_for_command,
@@ -31,7 +26,6 @@ from proliferate.server.cloud.commands.domain.rules import (
     validate_command_shape,
     validate_command_source,
 )
-from proliferate.server.cloud.commands.domain.target import target_requires_cloud_workspace
 from proliferate.server.cloud.commands.models import CreateCloudCommandRequest
 from proliferate.server.cloud.commands.preflight import (
     populate_agent_auth_preflight_payload,
@@ -41,6 +35,11 @@ from proliferate.server.cloud.commands.preflight import (
     validate_runtime_config_preflight,
 )
 from proliferate.server.cloud.commands.wake import kick_off_command_wake_after_commit_if_required
+from proliferate.server.cloud.commands.workspace_scope import (
+    command_has_managed_cloud_workspace,
+    direct_start_session_workspace_id,
+    resolve_command_workspace,
+)
 from proliferate.server.cloud.errors import CloudApiError
 from proliferate.server.cloud.event_logging import log_cloud_event
 from proliferate.server.cloud.live.service import publish_command_status_after_commit
@@ -65,660 +64,6 @@ def _idempotency_scope_for_command(
         f"user:{target.owner_user_id}:target:{target.id}:"
         f"workspace:{workspace_scope}:session:{session_scope}:kind:{kind}"
     )
-
-
-def _str_or_none(value: object) -> str | None:
-    return value if isinstance(value, str) and value else None
-
-
-_PROJECTED_SESSION_COMMAND_KINDS = {
-    CloudCommandKind.send_prompt.value,
-    CloudCommandKind.decide_plan.value,
-    CloudCommandKind.resolve_interaction.value,
-    CloudCommandKind.update_session_config.value,
-    CloudCommandKind.cancel_turn.value,
-    CloudCommandKind.close_session.value,
-}
-
-
-async def _resolve_command_workspace(
-    db: AsyncSession,
-    *,
-    user: ActorIdentity,
-    target: targets_store.CloudTargetSnapshot,
-    kind: str,
-    body: CreateCloudCommandRequest,
-) -> tuple[str | None, dict[str, object], str | None]:
-    if kind == CloudCommandKind.materialize_workspace.value and target_requires_cloud_workspace(
-        target
-    ):
-        if body.cloud_workspace_id is None:
-            raise CloudApiError(
-                "cloud_command_cloud_workspace_required",
-                "Managed materialize_workspace commands require cloudWorkspaceId.",
-                status_code=400,
-            )
-        cloud_workspace_id = await _resolve_cloud_workspace_id_for_target(
-            db,
-            target=target,
-            cloud_workspace_id=body.cloud_workspace_id,
-        )
-        exposure = await exposures_store.get_active_workspace_exposure(
-            db,
-            target_id=target.id,
-            cloud_workspace_id=UUID(cloud_workspace_id),
-        )
-        if exposure is None or exposure.archived_at is not None or exposure.status != "active":
-            raise CloudApiError(
-                "cloud_command_exposure_not_active",
-                "Workspace is not exposed for Cloud materialization.",
-                status_code=409,
-            )
-        workspace = await cloud_workspaces.get_cloud_workspace_by_id(
-            db,
-            UUID(cloud_workspace_id),
-        )
-        can_rematerialize_pruned_workspace = (
-            workspace is not None
-            and workspace.archived_at is None
-            and workspace.anyharness_workspace_id is None
-            and workspace.status == CloudWorkspaceStatus.needs_rematerialization.value
-        )
-        if not exposure.commandable and not can_rematerialize_pruned_workspace:
-            raise CloudApiError(
-                "cloud_command_exposure_not_commandable",
-                "Workspace exposure is read-only.",
-                status_code=409,
-            )
-        await require_workspace_interact(
-            db,
-            actor_user_id=user.id,
-            owner_scope=exposure.owner_scope,
-            owner_user_id=exposure.owner_user_id,
-            organization_id=exposure.organization_id,
-            workspace_archived=False,
-            exposure=exposure,
-        )
-        return None, body.payload, cloud_workspace_id
-    if kind in _PROJECTED_SESSION_COMMAND_KINDS:
-        return await _resolve_projected_session_command_workspace(
-            db,
-            user=user,
-            target=target,
-            body=body,
-        )
-    if kind == CloudCommandKind.backfill_exposed_workspace.value:
-        return await _resolve_backfill_exposed_workspace_command(
-            db,
-            user=user,
-            target=target,
-            body=body,
-        )
-    if kind == CloudCommandKind.prune_workspace_worktree.value:
-        return await _resolve_prune_workspace_worktree_command(
-            db,
-            user=user,
-            target=target,
-            body=body,
-        )
-    if kind == CloudCommandKind.start_session.value and target_requires_cloud_workspace(target):
-        if body.cloud_workspace_id is None and not body.workspace_id:
-            raise CloudApiError(
-                "cloud_command_cloud_workspace_required",
-                "Managed start_session commands require cloudWorkspaceId.",
-                status_code=400,
-            )
-        return await _resolve_managed_start_session_workspace(
-            db,
-            user=user,
-            target=target,
-            body=body,
-        )
-    if kind != CloudCommandKind.start_session.value:
-        cloud_workspace_id = None
-        if body.cloud_workspace_id is not None:
-            cloud_workspace_id = await _resolve_cloud_workspace_id_for_target(
-                db,
-                target=target,
-                cloud_workspace_id=body.cloud_workspace_id,
-            )
-        return body.workspace_id, body.payload, cloud_workspace_id
-    if not body.workspace_id:
-        workspace_id = _direct_start_session_workspace_id(body.payload)
-        if workspace_id is None:
-            return None, body.payload, None
-        return await _resolve_direct_start_session_workspace(
-            db,
-            user=user,
-            target=target,
-            body=body,
-            workspace_id=workspace_id,
-        )
-    try:
-        cloud_workspace_id = UUID(body.workspace_id or "")
-    except ValueError as exc:
-        raise CloudApiError(
-            "cloud_command_workspace_not_found",
-            "Start-session commands must reference a cloud workspace.",
-            status_code=404,
-        ) from exc
-
-    workspace = await cloud_workspaces.get_cloud_workspace_for_user(
-        db,
-        user.id,
-        cloud_workspace_id,
-    )
-    if workspace is None:
-        raise CloudApiError(
-            "cloud_command_workspace_not_found",
-            "Workspace not found.",
-            status_code=404,
-        )
-    if (
-        workspace.status != CloudWorkspaceStatus.ready.value
-        or not workspace.anyharness_workspace_id
-    ):
-        raise CloudApiError(
-            "cloud_command_workspace_not_ready",
-            "Workspace is not ready for cloud commands.",
-            status_code=409,
-        )
-    if workspace.target_id is not None:
-        if workspace.target_id != target.id:
-            raise CloudApiError(
-                "cloud_command_workspace_target_mismatch",
-                "Workspace is not attached to the requested target.",
-                status_code=409,
-            )
-    else:
-        runtime_environment = (
-            await cloud_runtime_environments.get_runtime_environment_for_workspace(
-                db,
-                workspace,
-            )
-        )
-        if runtime_environment is None or runtime_environment.target_id != target.id:
-            raise CloudApiError(
-                "cloud_command_workspace_target_mismatch",
-                "Workspace is not attached to the requested target.",
-                status_code=409,
-            )
-    exposure = await exposures_store.get_active_workspace_exposure(
-        db,
-        target_id=target.id,
-        cloud_workspace_id=workspace.id,
-    )
-    if (
-        exposure is None
-        or exposure.archived_at is not None
-        or exposure.status != "active"
-        or not exposure.anyharness_workspace_id
-    ):
-        raise CloudApiError(
-            "cloud_command_exposure_not_active",
-            "Workspace is not exposed for Cloud commands.",
-            status_code=409,
-        )
-    if not exposure.commandable:
-        raise CloudApiError(
-            "cloud_command_exposure_not_commandable",
-            "Workspace exposure is read-only.",
-            status_code=409,
-        )
-    await require_workspace_interact(
-        db,
-        actor_user_id=user.id,
-        owner_scope=exposure.owner_scope,
-        owner_user_id=exposure.owner_user_id,
-        organization_id=exposure.organization_id,
-        workspace_archived=workspace.archived_at is not None,
-        exposure=exposure,
-    )
-    payload = dict(body.payload)
-    payload["workspaceId"] = exposure.anyharness_workspace_id
-    return exposure.anyharness_workspace_id, payload, str(workspace.id)
-
-
-async def _resolve_direct_start_session_workspace(
-    db: AsyncSession,
-    *,
-    user: ActorIdentity,
-    target: targets_store.CloudTargetSnapshot,
-    body: CreateCloudCommandRequest,
-    workspace_id: str,
-) -> tuple[str | None, dict[str, object], str | None]:
-    cloud_workspace_id = await events_store.resolve_cloud_workspace_id(
-        db,
-        target_id=target.id,
-        workspace_id=workspace_id,
-    )
-    if cloud_workspace_id is None:
-        raise CloudApiError(
-            "cloud_command_exposure_not_active",
-            "Workspace is not exposed for Cloud commands.",
-            status_code=409,
-        )
-    exposure = await exposures_store.get_active_workspace_exposure(
-        db,
-        target_id=target.id,
-        cloud_workspace_id=cloud_workspace_id,
-    )
-    if (
-        exposure is None
-        or exposure.archived_at is not None
-        or exposure.status != "active"
-        or not exposure.anyharness_workspace_id
-    ):
-        raise CloudApiError(
-            "cloud_command_exposure_not_active",
-            "Workspace is not exposed for Cloud commands.",
-            status_code=409,
-        )
-    if not exposure.commandable:
-        raise CloudApiError(
-            "cloud_command_exposure_not_commandable",
-            "Workspace exposure is read-only.",
-            status_code=409,
-        )
-    await require_workspace_interact(
-        db,
-        actor_user_id=user.id,
-        owner_scope=exposure.owner_scope,
-        owner_user_id=exposure.owner_user_id,
-        organization_id=exposure.organization_id,
-        workspace_archived=False,
-        exposure=exposure,
-    )
-    payload = dict(body.payload)
-    payload["workspaceId"] = exposure.anyharness_workspace_id
-    return exposure.anyharness_workspace_id, payload, str(cloud_workspace_id)
-
-
-async def _resolve_managed_start_session_workspace(
-    db: AsyncSession,
-    *,
-    user: ActorIdentity,
-    target: targets_store.CloudTargetSnapshot,
-    body: CreateCloudCommandRequest,
-) -> tuple[str | None, dict[str, object], str | None]:
-    cloud_workspace_id = body.cloud_workspace_id
-    if cloud_workspace_id is None:
-        try:
-            cloud_workspace_id = UUID(body.workspace_id or "")
-        except ValueError as exc:
-            raise CloudApiError(
-                "cloud_command_workspace_not_found",
-                "Managed start-session commands must reference a Cloud workspace.",
-                status_code=404,
-            ) from exc
-
-    workspace = await cloud_workspaces.get_cloud_workspace_by_id(db, cloud_workspace_id)
-    if workspace is None or workspace.archived_at is not None:
-        raise CloudApiError(
-            "cloud_command_workspace_not_found",
-            "Workspace not found.",
-            status_code=404,
-        )
-    if workspace.status != CloudWorkspaceStatus.ready.value:
-        raise CloudApiError(
-            "cloud_command_workspace_not_ready",
-            "Workspace is not ready for cloud commands.",
-            status_code=409,
-        )
-    if (
-        workspace.target_id != target.id
-        or workspace.sandbox_profile_id != target.sandbox_profile_id
-        or workspace.owner_scope != target.owner_scope
-        or workspace.owner_user_id != target.owner_user_id
-        or workspace.organization_id != target.organization_id
-    ):
-        raise CloudApiError(
-            "cloud_command_workspace_target_mismatch",
-            "Workspace is not attached to the requested target.",
-            status_code=409,
-        )
-    if workspace.sandbox_profile_id is None or workspace.target_id is None:
-        raise CloudApiError(
-            "cloud_command_workspace_target_missing",
-            "Workspace is missing its managed sandbox profile target.",
-            status_code=409,
-        )
-    if workspace.materialized_target_id != target.id:
-        raise CloudApiError(
-            "cloud_command_workspace_target_stale",
-            "Workspace must be rematerialized on the active managed sandbox before commands run.",
-            status_code=409,
-        )
-    exposure = await exposures_store.get_active_workspace_exposure(
-        db,
-        target_id=target.id,
-        cloud_workspace_id=workspace.id,
-    )
-    if (
-        exposure is None
-        or exposure.archived_at is not None
-        or exposure.status != "active"
-        or not exposure.anyharness_workspace_id
-    ):
-        raise CloudApiError(
-            "cloud_command_exposure_not_active",
-            "Workspace is not exposed for Cloud commands.",
-            status_code=409,
-        )
-    if not exposure.commandable:
-        raise CloudApiError(
-            "cloud_command_exposure_not_commandable",
-            "Workspace exposure is read-only.",
-            status_code=409,
-        )
-    await require_workspace_interact(
-        db,
-        actor_user_id=user.id,
-        owner_scope=workspace.owner_scope,
-        owner_user_id=workspace.owner_user_id,
-        organization_id=workspace.organization_id,
-        workspace_archived=workspace.archived_at is not None,
-        exposure=exposure,
-    )
-    payload = dict(body.payload)
-    payload["workspaceId"] = exposure.anyharness_workspace_id
-    return exposure.anyharness_workspace_id, payload, str(workspace.id)
-
-
-async def _resolve_backfill_exposed_workspace_command(
-    db: AsyncSession,
-    *,
-    user: ActorIdentity,
-    target: targets_store.CloudTargetSnapshot,
-    body: CreateCloudCommandRequest,
-) -> tuple[str | None, dict[str, object], str | None]:
-    if body.cloud_workspace_id is None:
-        raise CloudApiError(
-            "cloud_command_cloud_workspace_required",
-            "backfill_exposed_workspace commands require cloudWorkspaceId.",
-            status_code=400,
-        )
-    cloud_workspace_id = await _resolve_cloud_workspace_id_for_target(
-        db,
-        target=target,
-        cloud_workspace_id=body.cloud_workspace_id,
-    )
-    exposure = await exposures_store.get_active_workspace_exposure(
-        db,
-        target_id=target.id,
-        cloud_workspace_id=UUID(cloud_workspace_id),
-    )
-    if (
-        exposure is None
-        or exposure.archived_at is not None
-        or exposure.status != "active"
-        or not exposure.anyharness_workspace_id
-    ):
-        raise CloudApiError(
-            "cloud_command_exposure_not_active",
-            "Workspace is not exposed for Cloud backfill.",
-            status_code=409,
-        )
-    if body.workspace_id and body.workspace_id != exposure.anyharness_workspace_id:
-        raise CloudApiError(
-            "cloud_command_workspace_target_mismatch",
-            "Backfill workspace id does not match the active exposure.",
-            status_code=409,
-        )
-    await require_workspace_interact(
-        db,
-        actor_user_id=user.id,
-        owner_scope=exposure.owner_scope,
-        owner_user_id=exposure.owner_user_id,
-        organization_id=exposure.organization_id,
-        workspace_archived=False,
-        exposure=exposure,
-    )
-    payload = dict(body.payload)
-    payload["workspaceId"] = exposure.anyharness_workspace_id
-    return exposure.anyharness_workspace_id, payload, cloud_workspace_id
-
-
-async def _resolve_prune_workspace_worktree_command(
-    db: AsyncSession,
-    *,
-    user: ActorIdentity,
-    target: targets_store.CloudTargetSnapshot,
-    body: CreateCloudCommandRequest,
-) -> tuple[str | None, dict[str, object], str | None]:
-    if body.cloud_workspace_id is None:
-        raise CloudApiError(
-            "cloud_command_cloud_workspace_required",
-            "prune_workspace_worktree commands require cloudWorkspaceId.",
-            status_code=400,
-        )
-    if not body.workspace_id:
-        raise CloudApiError(
-            "cloud_command_workspace_required",
-            "prune_workspace_worktree commands require workspaceId.",
-            status_code=400,
-        )
-    workspace = await cloud_workspaces.get_cloud_workspace_by_id(db, body.cloud_workspace_id)
-    if workspace is None:
-        raise CloudApiError(
-            "cloud_command_workspace_not_found",
-            "Workspace not found.",
-            status_code=404,
-        )
-    if (
-        workspace.target_id != target.id
-        or workspace.sandbox_profile_id != target.sandbox_profile_id
-        or workspace.owner_scope != target.owner_scope
-        or workspace.owner_user_id != target.owner_user_id
-        or workspace.organization_id != target.organization_id
-    ):
-        raise CloudApiError(
-            "cloud_command_workspace_target_mismatch",
-            "Workspace is not attached to the requested target.",
-            status_code=409,
-        )
-    if workspace.anyharness_workspace_id != body.workspace_id:
-        raise CloudApiError(
-            "cloud_command_workspace_target_mismatch",
-            "Prune workspace id does not match the Cloud workspace materialization.",
-            status_code=409,
-        )
-    exposure = await exposures_store.get_active_workspace_exposure(
-        db,
-        target_id=target.id,
-        cloud_workspace_id=workspace.id,
-    )
-    if (
-        exposure is not None
-        and exposure.anyharness_workspace_id is not None
-        and exposure.anyharness_workspace_id != body.workspace_id
-    ):
-        raise CloudApiError(
-            "cloud_command_workspace_target_mismatch",
-            "Prune workspace id does not match the active exposure.",
-            status_code=409,
-        )
-    await require_workspace_interact(
-        db,
-        actor_user_id=user.id,
-        owner_scope=workspace.owner_scope,
-        owner_user_id=workspace.owner_user_id,
-        organization_id=workspace.organization_id,
-        workspace_archived=workspace.archived_at is not None,
-        exposure=exposure,
-    )
-    payload = dict(body.payload)
-    payload["workspaceId"] = body.workspace_id
-    payload["cloudWorkspaceId"] = str(workspace.id)
-    return body.workspace_id, payload, str(workspace.id)
-
-
-async def _resolve_projected_session_command_workspace(
-    db: AsyncSession,
-    *,
-    user: ActorIdentity,
-    target: targets_store.CloudTargetSnapshot,
-    body: CreateCloudCommandRequest,
-) -> tuple[str | None, dict[str, object], str | None]:
-    if not body.session_id:
-        raise CloudApiError(
-            "cloud_command_session_required",
-            "Projected session commands require sessionId.",
-            status_code=400,
-        )
-    projection = await events_store.get_session_projection(
-        db,
-        target_id=target.id,
-        session_id=body.session_id,
-    )
-    if projection is None or projection.cloud_workspace_id is None:
-        raise CloudApiError(
-            "cloud_command_session_not_projected",
-            "Session is not projected into Cloud.",
-            status_code=409,
-        )
-    if (
-        body.cloud_workspace_id is not None
-        and body.cloud_workspace_id != projection.cloud_workspace_id
-    ):
-        raise CloudApiError(
-            "cloud_command_workspace_target_mismatch",
-            "Session is not attached to the requested Cloud workspace.",
-            status_code=409,
-        )
-    exposure = None
-    if projection.exposure_id is not None:
-        exposure = await exposures_store.get_workspace_exposure_by_id(
-            db,
-            projection.exposure_id,
-        )
-    if exposure is None and projection.cloud_workspace_id is not None:
-        exposure = await exposures_store.get_active_workspace_exposure(
-            db,
-            target_id=target.id,
-            cloud_workspace_id=projection.cloud_workspace_id,
-        )
-    if exposure is None or exposure.archived_at is not None or exposure.status != "active":
-        raise CloudApiError(
-            "cloud_command_exposure_not_active",
-            "Session is not exposed for Cloud commands.",
-            status_code=409,
-        )
-    if not exposure.commandable or not projection.commandable:
-        raise CloudApiError(
-            "cloud_command_exposure_not_commandable",
-            "Session exposure is read-only.",
-            status_code=409,
-        )
-    await require_workspace_interact(
-        db,
-        actor_user_id=user.id,
-        owner_scope=exposure.owner_scope,
-        owner_user_id=exposure.owner_user_id,
-        organization_id=exposure.organization_id,
-        workspace_archived=False,
-        exposure=exposure,
-    )
-    if body.kind == CloudCommandKind.decide_plan.value:
-        resolved_workspace_id = projection.workspace_id or exposure.anyharness_workspace_id
-    else:
-        resolved_workspace_id = (
-            projection.workspace_id or exposure.anyharness_workspace_id or body.workspace_id
-        )
-    if (
-        body.workspace_id is not None
-        and resolved_workspace_id is not None
-        and body.workspace_id != resolved_workspace_id
-    ):
-        raise CloudApiError(
-            "cloud_command_workspace_target_mismatch",
-            "Session is not attached to the requested runtime workspace.",
-            status_code=409,
-        )
-    payload = body.payload
-    if body.kind == CloudCommandKind.decide_plan.value:
-        if resolved_workspace_id is None:
-            raise CloudApiError(
-                "cloud_command_workspace_required",
-                "Projected plan decisions require a runtime workspace id.",
-                status_code=409,
-            )
-        payload = dict(body.payload)
-        payload_workspace_id = _str_or_none(payload.get("workspaceId"))
-        if payload_workspace_id is not None and payload_workspace_id != resolved_workspace_id:
-            raise CloudApiError(
-                "cloud_command_workspace_target_mismatch",
-                "Plan decision workspace does not match the projected session workspace.",
-                status_code=409,
-            )
-        payload["workspaceId"] = resolved_workspace_id
-    return (
-        resolved_workspace_id,
-        payload,
-        str(projection.cloud_workspace_id),
-    )
-
-
-async def _resolve_cloud_workspace_id_for_target(
-    db: AsyncSession,
-    *,
-    target: targets_store.CloudTargetSnapshot,
-    cloud_workspace_id: UUID,
-) -> str:
-    workspace = await cloud_workspaces.get_cloud_workspace_by_id(
-        db,
-        cloud_workspace_id,
-    )
-    if workspace is None or workspace.archived_at is not None:
-        raise CloudApiError(
-            "cloud_command_workspace_not_found",
-            "Workspace not found.",
-            status_code=404,
-        )
-    requires_managed_workspace = target_requires_cloud_workspace(target)
-    if (
-        workspace.target_id != target.id
-        or (
-            requires_managed_workspace
-            and workspace.sandbox_profile_id != target.sandbox_profile_id
-        )
-        or workspace.owner_scope != target.owner_scope
-        or workspace.owner_user_id != target.owner_user_id
-        or workspace.organization_id != target.organization_id
-    ):
-        raise CloudApiError(
-            "cloud_command_workspace_target_mismatch",
-            "Workspace is not attached to the requested target.",
-            status_code=409,
-        )
-    return str(workspace.id)
-
-
-def _direct_start_session_workspace_id(payload: dict[str, object]) -> str | None:
-    value = payload.get("workspaceId")
-    if not isinstance(value, str) or not value.strip():
-        return None
-    return value.strip()
-
-
-def _command_has_managed_cloud_workspace(
-    *,
-    target: targets_store.CloudTargetSnapshot,
-    kind: str,
-    body: CreateCloudCommandRequest,
-) -> bool:
-    if not target_requires_cloud_workspace(target):
-        return False
-    if kind == CloudCommandKind.start_session.value:
-        return body.cloud_workspace_id is not None or body.workspace_id is not None
-    if kind in {
-        CloudCommandKind.backfill_exposed_workspace.value,
-        CloudCommandKind.prune_workspace_worktree.value,
-    }:
-        return body.cloud_workspace_id is not None
-    if kind in _PROJECTED_SESSION_COMMAND_KINDS:
-        return body.cloud_workspace_id is not None
-    return False
 
 
 async def enqueue_command(
@@ -766,7 +111,7 @@ async def enqueue_command(
                 and body.cloud_workspace_id is not None
                 else None
             )
-            or _direct_start_session_workspace_id(body.payload)
+            or direct_start_session_workspace_id(body.payload)
         ),
         session_id=body.session_id,
         preconditions=body.preconditions,
@@ -790,7 +135,7 @@ async def enqueue_command(
         target=target,
         kind=kind,
         payload=payload,
-        require_target_config=not _command_has_managed_cloud_workspace(
+        require_target_config=not command_has_managed_cloud_workspace(
             target=target,
             kind=kind,
             body=body,
@@ -808,14 +153,14 @@ async def enqueue_command(
         actor_user_id=user.id,
         target=target,
         kind=kind,
-        require_target_config=not _command_has_managed_cloud_workspace(
+        require_target_config=not command_has_managed_cloud_workspace(
             target=target,
             kind=kind,
             body=body,
         ),
     )
     command_body = body.model_copy(update={"payload": payload})
-    resolved_workspace_id, payload, cloud_workspace_id = await _resolve_command_workspace(
+    resolved_workspace_id, payload, cloud_workspace_id = await resolve_command_workspace(
         db,
         user=user,
         target=target,
